Refactor TFRecordDatasetOp
